LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/transition-scrollbars-expected.txt

Foo bar foo bar foo Foo bar foo bar foo Foo bar foo bar foo Foo bar foo bar foo Foo bar foo bar foo

FAIL Scrollbars do not cause a transition of background-color assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 64, 128)"
